+++ src/test/groovy/org/liquibase/groovy/delegate/ColumnDelegateTests.groovy
@@ -15,7 +15,13 @@
 package org.liquibase.groovy.delegate
 
 import liquibase.change.AddColumnConfig
+import liquibase.change.Change
 import liquibase.change.ColumnConfig
+import liquibase.change.core.AddColumnChange
+import liquibase.change.core.CreateTableChange
+import liquibase.change.core.DeleteDataChange
+import liquibase.change.core.LoadDataChange
+import liquibase.change.core.UpdateDataChange
 import liquibase.exception.ChangeLogParseException
 import liquibase.statement.DatabaseFunction
 import liquibase.statement.SequenceCurrentValueFunction
@@ -32,6 +38,10 @@ import java.text.SimpleDateFormat
  * Test class for the {@link ColumnDelegate}.  As usual, we're only verifying that we can pass
  * things to Liquibase correctly. We check all attributes that are known at this time - note that
  * several are undocumented.
+ * <p>
+ * The columnDelegate also processes "where" and "whereParam" blocks.  For the most tests, we know
+ * nothing got set because we use Liquibase's CreateTableChange, which will throw an exception if
+ * we try to set values for "where" and "whereParam" erroneously.
  *
  * @author Steven C. Saliman
  */
@@ -44,40 +54,43 @@ class ColumnDelegateTests {
      */
     @Test
     void oneColumnEmptyNoClosure() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new CreateTableChange(), ColumnConfig.class, ) {
             column([:])
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertNull delegate.columns[0].name
-        assertNull delegate.columns[0].computed
-        assertNull delegate.columns[0].type
-        assertNull delegate.columns[0].value
-        assertNull delegate.columns[0].valueNumeric
-        assertNull delegate.columns[0].valueBoolean
-        assertNull delegate.columns[0].valueDate
-        assertNull delegate.columns[0].valueComputed
-        assertNull delegate.columns[0].valueSequenceNext
-        assertNull delegate.columns[0].valueSequenceCurrent
-        assertNull delegate.columns[0].valueBlobFile
-        assertNull delegate.columns[0].valueClobFile
-        assertNull delegate.columns[0].defaultValue
-        assertNull delegate.columns[0].defaultValueNumeric
-        assertNull delegate.columns[0].defaultValueDate
-        assertNull delegate.columns[0].defaultValueBoolean
-        assertNull delegate.columns[0].defaultValueComputed
-        assertNull delegate.columns[0].defaultValueSequenceNext
-        assertNull delegate.columns[0].defaultValueConstraintName
-        assertNull delegate.columns[0].autoIncrement
-        assertNull delegate.columns[0].startWith
-        assertNull delegate.columns[0].incrementBy
-        assertNull delegate.columns[0].defaultOnNull
-        assertNull delegate.columns[0].generationType
-        assertNull delegate.columns[0].remarks
-        assertNull delegate.columns[0].descending
-        assertNull delegate.columns[0].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertNull column.name
+        assertNull column.computed
+        assertNull column.type
+        assertNull column.value
+        assertNull column.valueNumeric
+        assertNull column.valueBoolean
+        assertNull column.valueDate
+        assertNull column.valueComputed
+        assertNull column.valueSequenceNext
+        assertNull column.valueSequenceCurrent
+        assertNull column.valueBlobFile
+        assertNull column.valueClobFile
+        assertNull column.defaultValue
+        assertNull column.defaultValueNumeric
+        assertNull column.defaultValueDate
+        assertNull column.defaultValueBoolean
+        assertNull column.defaultValueComputed
+        assertNull column.defaultValueSequenceNext
+        assertNull column.defaultValueConstraintName
+        assertNull column.autoIncrement
+        assertNull column.startWith
+        assertNull column.incrementBy
+        assertNull column.defaultOnNull
+        assertNull column.generationType
+        assertNull column.remarks
+        assertNull column.descending
+        assertNull column.constraints
     }
 
     /**
@@ -87,40 +100,43 @@ class ColumnDelegateTests {
      */
     @Test
     void oneColumnEmptyWithClosure() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             column([:]) {}
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertNull delegate.columns[0].name
-        assertNull delegate.columns[0].computed
-        assertNull delegate.columns[0].type
-        assertNull delegate.columns[0].value
-        assertNull delegate.columns[0].valueNumeric
-        assertNull delegate.columns[0].valueBoolean
-        assertNull delegate.columns[0].valueDate
-        assertNull delegate.columns[0].valueComputed
-        assertNull delegate.columns[0].valueSequenceNext
-        assertNull delegate.columns[0].valueSequenceCurrent
-        assertNull delegate.columns[0].valueBlobFile
-        assertNull delegate.columns[0].valueClobFile
-        assertNull delegate.columns[0].defaultValue
-        assertNull delegate.columns[0].defaultValueNumeric
-        assertNull delegate.columns[0].defaultValueDate
-        assertNull delegate.columns[0].defaultValueBoolean
-        assertNull delegate.columns[0].defaultValueComputed
-        assertNull delegate.columns[0].defaultValueSequenceNext
-        assertNull delegate.columns[0].defaultValueConstraintName
-        assertNull delegate.columns[0].autoIncrement
-        assertNull delegate.columns[0].startWith
-        assertNull delegate.columns[0].incrementBy
-        assertNull delegate.columns[0].defaultOnNull
-        assertNull delegate.columns[0].generationType
-        assertNull delegate.columns[0].remarks
-        assertNull delegate.columns[0].descending
-        assertNotNull delegate.columns[0].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertNull column.name
+        assertNull column.computed
+        assertNull column.type
+        assertNull column.value
+        assertNull column.valueNumeric
+        assertNull column.valueBoolean
+        assertNull column.valueDate
+        assertNull column.valueComputed
+        assertNull column.valueSequenceNext
+        assertNull column.valueSequenceCurrent
+        assertNull column.valueBlobFile
+        assertNull column.valueClobFile
+        assertNull column.defaultValue
+        assertNull column.defaultValueNumeric
+        assertNull column.defaultValueDate
+        assertNull column.defaultValueBoolean
+        assertNull column.defaultValueComputed
+        assertNull column.defaultValueSequenceNext
+        assertNull column.defaultValueConstraintName
+        assertNull column.autoIncrement
+        assertNull column.startWith
+        assertNull column.incrementBy
+        assertNull column.defaultOnNull
+        assertNull column.generationType
+        assertNull column.remarks
+        assertNull column.descending
+        assertNotNull column.constraints
     }
 
     /**
@@ -138,7 +154,7 @@ class ColumnDelegateTests {
         def defaultDate = "2013-12-31 09:30:04"
         def columnDefaultDate = parseSqlTimestamp(defaultDate)
 
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             column(
                     name: 'columnName',
                     computed: true,
@@ -169,36 +185,39 @@ class ColumnDelegateTests {
             )
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertEquals 'columnName', delegate.columns[0].name
-        assertTrue delegate.columns[0].computed
-        assertEquals 'timezonetz', delegate.columns[0].type
-        assertEquals 'someValue', delegate.columns[0].value
-        assertEquals 1, delegate.columns[0].valueNumeric.intValue()
-        assertFalse delegate.columns[0].valueBoolean
-        assertEquals columnDateValue, delegate.columns[0].valueDate
-        assertEquals 'databaseValue', delegate.columns[0].valueComputed.value
-        assertEquals 'sequenceNext', delegate.columns[0].valueSequenceNext.value
-        assertEquals 'sequenceCurrent', delegate.columns[0].valueSequenceCurrent.value
-        assertEquals 'someBlobFile', delegate.columns[0].valueBlobFile
-        assertEquals 'someClobFile', delegate.columns[0].valueClobFile
-        assertEquals 'someDefaultValue', delegate.columns[0].defaultValue
-        assertEquals 2, delegate.columns[0].defaultValueNumeric.intValue()
-        assertEquals columnDefaultDate, delegate.columns[0].defaultValueDate
-        assertFalse delegate.columns[0].defaultValueBoolean
-        assertEquals 'defaultDatabaseValue', delegate.columns[0].defaultValueComputed.value
-        assertEquals 'defaultSequence', delegate.columns[0].defaultValueSequenceNext.value
-        assertEquals 'defaultValueConstraint', delegate.columns[0].defaultValueConstraintName
-        assertTrue delegate.columns[0].autoIncrement
-        assertEquals 3G, delegate.columns[0].startWith
-        assertEquals 4G, delegate.columns[0].incrementBy
-        assertTrue delegate.columns[0].defaultOnNull
-        assertEquals 'someType', delegate.columns[0].generationType
-        assertEquals 'No comment', delegate.columns[0].remarks
-        assertTrue delegate.columns[0].descending
-        assertNull delegate.columns[0].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'columnName', column.name
+        assertTrue column.computed
+        assertEquals 'timezonetz', column.type
+        assertEquals 'someValue', column.value
+        assertEquals 1, column.valueNumeric.intValue()
+        assertFalse column.valueBoolean
+        assertEquals columnDateValue, column.valueDate
+        assertEquals 'databaseValue', column.valueComputed.value
+        assertEquals 'sequenceNext', column.valueSequenceNext.value
+        assertEquals 'sequenceCurrent', column.valueSequenceCurrent.value
+        assertEquals 'someBlobFile', column.valueBlobFile
+        assertEquals 'someClobFile', column.valueClobFile
+        assertEquals 'someDefaultValue', column.defaultValue
+        assertEquals 2, column.defaultValueNumeric.intValue()
+        assertEquals columnDefaultDate, column.defaultValueDate
+        assertFalse column.defaultValueBoolean
+        assertEquals 'defaultDatabaseValue', column.defaultValueComputed.value
+        assertEquals 'defaultSequence', column.defaultValueSequenceNext.value
+        assertEquals 'defaultValueConstraint', column.defaultValueConstraintName
+        assertTrue column.autoIncrement
+        assertEquals 3G, column.startWith
+        assertEquals 4G, column.incrementBy
+        assertTrue column.defaultOnNull
+        assertEquals 'someType', column.generationType
+        assertEquals 'No comment', column.remarks
+        assertTrue column.descending
+        assertNull column.constraints
     }
 
     /**
@@ -207,7 +226,7 @@ class ColumnDelegateTests {
      */
     @Test
     void twoColumns() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             // first one has only the boolean value set to true
             column(
                     name: 'first',
@@ -224,20 +243,25 @@ class ColumnDelegateTests {
             )
         }
 
-        assertNull delegate.whereClause
-        assertEquals 2, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertEquals 'first', delegate.columns[0].name
-        assertTrue delegate.columns[0].valueBoolean
-        assertFalse delegate.columns[0].defaultValueBoolean
-        assertFalse delegate.columns[0].autoIncrement
-        assertNull delegate.columns[0].constraints
-        assertTrue delegate.columns[1] instanceof ColumnConfig
-        assertEquals 'second', delegate.columns[1].name
-        assertFalse delegate.columns[1].valueBoolean
-        assertTrue delegate.columns[1].defaultValueBoolean
-        assertFalse delegate.columns[1].autoIncrement
-        assertNull delegate.columns[1].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 2, columns.size()
+
+        def column = columns[0]
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'first', column.name
+        assertTrue column.valueBoolean
+        assertFalse column.defaultValueBoolean
+        assertFalse column.autoIncrement
+        assertNull column.constraints
+
+        column = columns[1]
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'second', column.name
+        assertFalse column.valueBoolean
+        assertTrue column.defaultValueBoolean
+        assertFalse column.autoIncrement
+        assertNull column.constraints
 
     }
 
@@ -247,21 +271,24 @@ class ColumnDelegateTests {
      */
     @Test
     void columnWithConstraint() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             // first one has only the boolean value set to true
             column(name: 'first', type: 'int') {
                 constraints(nullable: false, unique: true)
             }
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertEquals 'first', delegate.columns[0].name
-        assertEquals 'int', delegate.columns[0].type
-        assertNotNull delegate.columns[0].constraints
-        assertFalse delegate.columns[0].constraints.nullable
-        assertTrue delegate.columns[0].constraints.unique
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'first', column.name
+        assertEquals 'int', column.type
+        assertNotNull column.constraints
+        assertFalse column.constraints.nullable
+        assertTrue column.constraints.unique
     }
 
     /**
@@ -280,7 +307,7 @@ class ColumnDelegateTests {
         def defaultDate = "2013-12-31 09:30:04"
         def columnDefaultDate = parseSqlTimestamp(defaultDate)
 
-        def delegate = buildColumnDelegate(AddColumnConfig.class) {
+        def delegate = buildColumnDelegate(new AddColumnChange(), AddColumnConfig.class) {
             column(
                     name: 'columnName',
                     computed: false,
@@ -314,39 +341,42 @@ class ColumnDelegateTests {
             )
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof AddColumnConfig
-        assertEquals 'columnName', delegate.columns[0].name
-        assertFalse delegate.columns[0].computed
-        assertEquals 'varchar(30)', delegate.columns[0].type
-        assertEquals 'someValue', delegate.columns[0].value
-        assertEquals 1, delegate.columns[0].valueNumeric.intValue()
-        assertFalse delegate.columns[0].valueBoolean
-        assertEquals columnDateValue, delegate.columns[0].valueDate
-        assertEquals 'databaseValue', delegate.columns[0].valueComputed.value
-        assertEquals 'sequenceNext', delegate.columns[0].valueSequenceNext.value
-        assertEquals 'sequenceCurrent', delegate.columns[0].valueSequenceCurrent.value
-        assertEquals 'someBlobFile', delegate.columns[0].valueBlobFile
-        assertEquals 'someClobFile', delegate.columns[0].valueClobFile
-        assertEquals 'someDefaultValue', delegate.columns[0].defaultValue
-        assertEquals 2, delegate.columns[0].defaultValueNumeric.intValue()
-        assertEquals columnDefaultDate, delegate.columns[0].defaultValueDate
-        assertFalse delegate.columns[0].defaultValueBoolean
-        assertEquals 'defaultDatabaseValue', delegate.columns[0].defaultValueComputed.value
-        assertEquals 'defaultSequence', delegate.columns[0].defaultValueSequenceNext.value
-        assertEquals 'defaultValueConstraint', delegate.columns[0].defaultValueConstraintName
-        assertTrue delegate.columns[0].autoIncrement
-        assertEquals 3G, delegate.columns[0].startWith
-        assertEquals 4G, delegate.columns[0].incrementBy
-        assertTrue delegate.columns[0].defaultOnNull
-        assertEquals 'someType', delegate.columns[0].generationType
-        assertEquals 'No comment', delegate.columns[0].remarks
-        assertFalse delegate.columns[0].descending
-        assertEquals 'before', delegate.columns[0].beforeColumn
-        assertEquals 'after', delegate.columns[0].afterColumn
-        assertEquals 5, delegate.columns[0].position
-        assertNull delegate.columns[0].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof AddColumnConfig
+        assertEquals 'columnName', column.name
+        assertFalse column.computed
+        assertEquals 'varchar(30)', column.type
+        assertEquals 'someValue', column.value
+        assertEquals 1, column.valueNumeric.intValue()
+        assertFalse column.valueBoolean
+        assertEquals columnDateValue, column.valueDate
+        assertEquals 'databaseValue', column.valueComputed.value
+        assertEquals 'sequenceNext', column.valueSequenceNext.value
+        assertEquals 'sequenceCurrent', column.valueSequenceCurrent.value
+        assertEquals 'someBlobFile', column.valueBlobFile
+        assertEquals 'someClobFile', column.valueClobFile
+        assertEquals 'someDefaultValue', column.defaultValue
+        assertEquals 2, column.defaultValueNumeric.intValue()
+        assertEquals columnDefaultDate, column.defaultValueDate
+        assertFalse column.defaultValueBoolean
+        assertEquals 'defaultDatabaseValue', column.defaultValueComputed.value
+        assertEquals 'defaultSequence', column.defaultValueSequenceNext.value
+        assertEquals 'defaultValueConstraint', column.defaultValueConstraintName
+        assertTrue column.autoIncrement
+        assertEquals 3G, column.startWith
+        assertEquals 4G, column.incrementBy
+        assertTrue column.defaultOnNull
+        assertEquals 'someType', column.generationType
+        assertEquals 'No comment', column.remarks
+        assertFalse column.descending
+        assertEquals 'before', column.beforeColumn
+        assertEquals 'after', column.afterColumn
+        assertEquals 5, column.position
+        assertNull column.constraints
     }
 
     /**
@@ -364,7 +394,7 @@ class ColumnDelegateTests {
         def defaultDate = "2013-12-31 09:30:04"
         def columnDefaultDate = parseSqlTimestamp(defaultDate)
 
-        def delegate = buildColumnDelegate(LoadDataColumnConfig.class) {
+        def delegate = buildColumnDelegate(new LoadDataChange(), LoadDataColumnConfig.class) {
             column(
                     name: 'columnName',
                     computed: true,
@@ -398,71 +428,164 @@ class ColumnDelegateTests {
             )
         }
 
-        assertNull delegate.whereClause
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof LoadDataColumnConfig
-        assertEquals 'columnName', delegate.columns[0].name
-        assertTrue delegate.columns[0].computed
-        assertEquals 'STRING', delegate.columns[0].type
-        assertEquals 'someValue', delegate.columns[0].value
-        assertEquals 1, delegate.columns[0].valueNumeric.intValue()
-        assertFalse delegate.columns[0].valueBoolean
-        assertEquals columnDateValue, delegate.columns[0].valueDate
-        assertEquals 'databaseValue', delegate.columns[0].valueComputed.value
-        assertEquals 'sequenceNext', delegate.columns[0].valueSequenceNext.value
-        assertEquals 'sequenceCurrent', delegate.columns[0].valueSequenceCurrent.value
-        assertEquals 'someBlobFile', delegate.columns[0].valueBlobFile
-        assertEquals 'someClobFile', delegate.columns[0].valueClobFile
-        assertEquals 'someDefaultValue', delegate.columns[0].defaultValue
-        assertEquals 2, delegate.columns[0].defaultValueNumeric.intValue()
-        assertEquals columnDefaultDate, delegate.columns[0].defaultValueDate
-        assertFalse delegate.columns[0].defaultValueBoolean
-        assertEquals 'defaultDatabaseValue', delegate.columns[0].defaultValueComputed.value
-        assertEquals 'defaultSequence', delegate.columns[0].defaultValueSequenceNext.value
-        assertEquals 'defaultValueConstraint', delegate.columns[0].defaultValueConstraintName
-        assertTrue delegate.columns[0].autoIncrement
-        assertEquals 3G, delegate.columns[0].startWith
-        assertEquals 4G, delegate.columns[0].incrementBy
-        assertTrue delegate.columns[0].defaultOnNull
-        assertEquals 'someType', delegate.columns[0].generationType
-        assertEquals 'No comment', delegate.columns[0].remarks
-        assertFalse delegate.columns[0].descending
-        assertEquals 'columnHeader', delegate.columns[0].header
-        assertEquals 5, delegate.columns[0].index
-        assertFalse delegate.columns[0].allowUpdate
-        assertNull delegate.columns[0].constraints
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof LoadDataColumnConfig
+        assertEquals 'columnName', column.name
+        assertTrue column.computed
+        assertEquals 'STRING', column.type
+        assertEquals 'someValue', column.value
+        assertEquals 1, column.valueNumeric.intValue()
+        assertFalse column.valueBoolean
+        assertEquals columnDateValue, column.valueDate
+        assertEquals 'databaseValue', column.valueComputed.value
+        assertEquals 'sequenceNext', column.valueSequenceNext.value
+        assertEquals 'sequenceCurrent', column.valueSequenceCurrent.value
+        assertEquals 'someBlobFile', column.valueBlobFile
+        assertEquals 'someClobFile', column.valueClobFile
+        assertEquals 'someDefaultValue', column.defaultValue
+        assertEquals 2, column.defaultValueNumeric.intValue()
+        assertEquals columnDefaultDate, column.defaultValueDate
+        assertFalse column.defaultValueBoolean
+        assertEquals 'defaultDatabaseValue', column.defaultValueComputed.value
+        assertEquals 'defaultSequence', column.defaultValueSequenceNext.value
+        assertEquals 'defaultValueConstraint', column.defaultValueConstraintName
+        assertTrue column.autoIncrement
+        assertEquals 3G, column.startWith
+        assertEquals 4G, column.incrementBy
+        assertTrue column.defaultOnNull
+        assertEquals 'someType', column.generationType
+        assertEquals 'No comment', column.remarks
+        assertFalse column.descending
+        assertEquals 'columnHeader', column.header
+        assertEquals 5, column.index
+        assertFalse column.allowUpdate
+        assertNull column.constraints
     }
 
     /**
-     * Test a column closure that has a where clause.
+     * Test a column closure that has a where clause.  For this test, we need a class that supports
+     * where clauses.
      */
     @Test
     void columnClosureCanContainWhereClause() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        def delegate = buildColumnDelegate(new UpdateDataChange(), ColumnConfig.class) {
             column(name: 'monkey', type: 'VARCHAR(50)')
             where "emotion='angry'"
         }
 
-        assertNotNull delegate.columns
-        assertEquals 1, delegate.columns.size()
-        assertTrue delegate.columns[0] instanceof ColumnConfig
-        assertEquals 'monkey', delegate.columns[0].name
-        assertEquals "emotion='angry'", delegate.whereClause
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'monkey', column.name
+        assertEquals "emotion='angry'", change.where
+        assertEquals 0, change.whereParams.size()
+    }
+
+    /**
+     * Test a column closure that has a where clause.  For this test, we need a class that supports
+     * where clauses.  This test is the same as the last one, except we'll add some whereParams to
+     * the mix.
+     */
+    @Test
+    void columnClosureCanContainWhereClauseWithParams() {
+        def delegate = buildColumnDelegate(new UpdateDataChange(), ColumnConfig.class) {
+            column(name: 'monkey', type: 'VARCHAR(50)')
+            where "emotion=':emotion' and day=':monday'"
+            whereParams {
+                param(name: 'emotion', value: 'angry')
+                param(name: 'day', value: 'monday')
+            }
+        }
+
+        def change = delegate.change
+        def columns = change.columns
+        assertEquals 1, columns.size()
+        def column = columns[0]
+
+        assertTrue column instanceof ColumnConfig
+        assertEquals 'monkey', column.name
+        assertEquals "emotion=':emotion' and day=':monday'", change.where
+
+        assertEquals 2, change.whereParams.size()
+
+        column = change.whereParams[0]
+        assertEquals "emotion", column.name
+        assertEquals "angry", column.value
     }
 
     /**
      * {@code delete} changes will have a where clause, but no actual columns.  Make sure we can
-     * handle this.
+     * handle this.  We'll also use this test to put every documented attribute of a whereParam to
+     * make sure we handle them all
      */
     @Test
-    void columnClosureIsJustWhereClause() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+    void columnClosureIsJustWhereClauseWithParams() {
+        def dateValue = "2010-11-02 07:52:04"
+        def columnDateValue = parseSqlTimestamp(dateValue)
+
+        def delegate = buildColumnDelegate(new DeleteDataChange(), ColumnConfig.class) {
+            where "emotion=':emotion'"
+            whereParams{
+                param(name: 'emotion',
+                      value: 'angry',
+                      valueNumeric: 1,
+                      valueBoolean: false,
+                      valueDate: dateValue,
+                      valueComputed: new DatabaseFunction('databaseValue'),
+                      valueSequenceNext: new SequenceNextValueFunction('sequenceNext'),
+                      valueSequenceCurrent: new SequenceCurrentValueFunction('sequenceCurrent')
+                )
+            }
+        }
+
+        def change = delegate.change
+        assertEquals "emotion=':emotion'", change.where
+        assertEquals 1, change.whereParams.size()
+
+        def column = change.whereParams[0]
+        assertEquals "emotion", column.name
+        assertEquals "angry", column.value
+        assertEquals 1, column.valueNumeric.intValue()
+        assertFalse column.valueBoolean
+        assertEquals columnDateValue, column.valueDate
+        assertEquals 'databaseValue', column.valueComputed.value
+        assertEquals 'sequenceNext', column.valueSequenceNext.value
+        assertEquals 'sequenceCurrent', column.valueSequenceCurrent.value
+    }
+
+    /**
+     * Try using a "where" clause in a change that doesn't support them, like the CreateTableChange.
+     * Expect a ChangeLogParseException.
+     */
+    @Test(expected = ChangeLogParseException)
+    void invalidUseOfWhere() {
+        buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
+            column(name: 'monkey', type: 'VARCHAR(50)')
             where "emotion='angry'"
         }
 
-        assertNotNull delegate.columns
-        assertEquals 0, delegate.columns.size()
-        assertEquals "emotion='angry'", delegate.whereClause
+    }
+
+    /**
+     * Try using a "whereParams" clause in a change that doesn't support them, like the
+     * CreateTableChange. Expect a ChangeLogParseException.
+     */
+    @Test(expected = ChangeLogParseException)
+    void invalidUseOfWhereParams() {
+        buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
+            column(name: 'monkey', type: 'VARCHAR(50)')
+            whereParams {
+                param(name: 'emotion', value: 'angry')
+            }
+        }
+
     }
 
     /**
@@ -471,15 +594,10 @@ class ColumnDelegateTests {
      */
     @Test(expected = ChangeLogParseException)
     void invalidMethodInClosure() {
-        def delegate = buildColumnDelegate(ColumnConfig.class) {
+        buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             table(name: 'monkey')
         }
-
-        assertNotNull delegate.columns
-        assertEquals 0, delegate.columns.size()
-        assertEquals "emotion='angry'", delegate.whereClause
     }
-
 
     /**
      * Try building a column when it contains an invalid attribute.  Do we get a
@@ -488,24 +606,27 @@ class ColumnDelegateTests {
      */
     @Test(expected = ChangeLogParseException)
     void columnWithInvalidAttribute() {
-        buildColumnDelegate(ColumnConfig.class) {
+        buildColumnDelegate(new CreateTableChange(), ColumnConfig.class) {
             column(header: 'invalid')
         }
     }
 
     /**
      * helper method to build and execute a ColumnDelegate.
-     * @param closure the closure to execute
+     * @param change the change that will be modified by the delegate.
+     * @param columnConfigClass the type of ColumnConfig used by the change.
+     * @param closure the closure to execute with our column attributes.
      * @return the new delegate.
      */
-    private def buildColumnDelegate(Class columnConfigClass, Closure closure) {
+    private def buildColumnDelegate(Change change, Class columnConfigClass, Closure closure) {
         def changelog = new DatabaseChangeLog()
         changelog.changeLogParameters = new ChangeLogParameters()
         def columnDelegate = new ColumnDelegate(
                 columnConfigClass: columnConfigClass,
                 databaseChangeLog: changelog,
                 changeSetId: 'test-change-set',
-                changeName: 'create-table'
+                changeName: 'test-change',
+                change: change
         )
         closure.delegate = columnDelegate
         closure.resolveStrategy = Closure.DELEGATE_FIRST